Implementing App Skeleton Commit #1
- Implemented the stack and tab navigator
- Implemented the sessions manager, sessions creation, data collection,
  and settings screens, with the session manager component structure
completed
