+++ screens/DataCollectionScreen.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { View, Text, Button } from "react-native";
+
+export default function DataCollectionScreen() {
+    return (
+        <View>
+            <Text> Data Collection </Text>
+        </View>
+    );
+}
+
