+++ screens/SessionsManagerScreen.tsx
@@ -1,9 +1,21 @@
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, Button, FlatList, Text, View } from 'react-native';
 
-export default function SessionsManagerScreen() {
+const data = new Array(100)
+  .fill(null)
+  .map((v, i) => ({ key: i.toString(), value: ('Item ' + i.toString())
+}));
+
+export default function SessionsManagerScreen({ route, navigation }) {
     return (
         <View>
-            <Text> Sessions </Text>
+            <FlatList
+                data={data}
+                renderItem={({item}) => <Button 
+                                            title={item.value} 
+                                            onPress={() => {navigation.navigate("Data Collection", {});}}
+                                        />
+                }
+            />
         </View>
     );
 }
